Elaborate on uninstalling package from system
Specify upfront that users most likely has to enable "Automatic Uninstaller" to remove a package from the system (which is most likely the expected/desired behavior when you uninstall a package via a package manager).

Explain the default behavior of `choco uninstall` (i.e. with "Automatic Uninstaller" disabled) in a later section.