remove cin.st
Even having the link there could cause security concerns
that we don't want to get into. It's better not to have
the link there at all because someone might see it and
decide to use it against better recommendations and bad
things *could* happen to their system that we have no
control over. Not that it would, but it's not outside
the realm of possibilities.