(layout) Hide uninstall scripts until button click
The uninstallation scripts are now hidden on Chocolatey.org unil a
button is clicked which reads "Yes, I understand the dangers of running
these scripts". This has been added as an extra layer of precaution to
ensure the user understands what might happen to their system if the
scripts are ran.