(doc) Update pictures in moderation docs
With the redesign of Chocolatey.org, some images needed to be updated
in the moderation docs. New images have been added to represent the
redesign and current look of the Community Package Moderation process.

Minor verbiage changes have been made to step 3 to include the changes
to the status ball that is next to each package title. Now instead of
representing only Verification testing, the ball also represents
Validation testing.