(maint) Correct link format
The link format in this commit was using the incorrect format that is
not used anywhere else in the repository. When clicking on the link, it
takes you to the raw file, and does not work at all on chocolatey.org.

The format has been corrected to how it used throughout the repository,
and now works on choco-wiki and on chocolatey.org.